ccp4 header size was wrong
git-svn-id: file:///home/roberto/Git/svn@7796 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 24a9c89bce4a9b5d502f7cff92dd697439c69e75

+++ tags/release-2.4/xmipp/libraries/interface/ccp4.h
@@ -122,7 +122,7 @@ typedef struct MRCheader
 
     int         nsymbt;             /* Number of bytes used for storing   */
     /* symmetry operators. */
-    unsigned long  user[MRC_USER];  /* For user - all set to zero by default. */
+    unsigned int  user[MRC_USER];  /* For user - all set to zero by default. */
     char        map[4];              /* Char string 'MAP ' to identify file type */
     char        machst[4];           /* Machine stamp indicating the machine type
                       which wrote file
